fixed problem in CriteriaTypeRepository.java. Changed all repositories.

+++ src/main/java/kg/attractor/xfood/repository/CheckListRepository.java
@@ -3,5 +3,5 @@ package kg.attractor.xfood.repository;
 import kg.attractor.xfood.model.CheckList;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CheckListRepository extends JpaRepository<CheckList, Integer> {
+public interface CheckListRepository extends JpaRepository<CheckList, Long> {
 }
