+++ src/main/java/kg/attractor/xfood/repository/OpportunityRepository.java
@@ -3,5 +3,5 @@ package kg.attractor.xfood.repository;
 import kg.attractor.xfood.model.Opportunity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OpportunityRepository extends JpaRepository<Opportunity, Integer> {
+public interface OpportunityRepository extends JpaRepository<Opportunity, Long> {
 }
