+++ src/main/java/kg/attractor/xfood/repository/SectionRepository.java
@@ -3,5 +3,5 @@ package kg.attractor.xfood.repository;
 import kg.attractor.xfood.model.Section;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SectionRepository extends JpaRepository<Section, Integer> {
+public interface SectionRepository extends JpaRepository<Section, Long> {
 }
